Move some classes from security/ee package
