+++ appserver/ejb/ejb-container/src/main/java/org/glassfish/ejb/security/application/EJBSecurityManager.java
@@ -23,9 +23,9 @@ import com.sun.enterprise.security.SecurityContext;
 import com.sun.enterprise.security.SecurityManager;
 import com.sun.enterprise.security.auth.login.LoginContextDriver;
 import com.sun.enterprise.security.authorize.PolicyContextHandlerImpl;
+import com.sun.enterprise.security.authorize.cache.PermissionCache;
+import com.sun.enterprise.security.authorize.cache.PermissionCacheFactory;
 import com.sun.enterprise.security.common.AppservAccessController;
-import com.sun.enterprise.security.ee.PermissionCache;
-import com.sun.enterprise.security.ee.PermissionCacheFactory;
 import com.sun.enterprise.security.ee.SecurityUtil;
 import com.sun.enterprise.security.ee.audit.AppServerAuditManager;
 import com.sun.logging.LogDomains;
