+++ appserver/security/core-ee/src/main/java/com/sun/enterprise/security/authorize/cache/CachedPermissionImpl.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package com.sun.enterprise.security.ee;
+package com.sun.enterprise.security.authorize.cache;
 
 import java.security.Permission;
