+++ appserver/security/core-ee/src/main/java/com/sun/enterprise/security/authorize/cache/PermissionCacheFactory.java
@@ -14,8 +14,9 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package com.sun.enterprise.security.ee;
+package com.sun.enterprise.security.authorize.cache;
 
+import com.sun.enterprise.security.ee.J2EESecurityManager;
 import java.security.CodeSource;
 import java.security.Permission;
 import java.security.Policy;
