+++ appserver/security/webservices.security/src/main/java/com/sun/enterprise/security/jmac/provider/config/PipeHelper.java
@@ -59,10 +59,10 @@ import com.sun.enterprise.security.SecurityContext;
 import com.sun.enterprise.security.SecurityServicesUtil;
 import com.sun.enterprise.security.appclient.ConfigXMLParser;
 import com.sun.enterprise.security.audit.AuditManager;
+import com.sun.enterprise.security.authorize.EJBPolicyContextDelegate;
 import com.sun.enterprise.security.common.AppservAccessController;
 import com.sun.enterprise.security.common.ClientSecurityContext;
 import com.sun.enterprise.security.ee.audit.AppServerAuditManager;
-import com.sun.enterprise.security.ee.authorize.EJBPolicyContextDelegate;
 import com.sun.enterprise.security.jmac.AuthMessagePolicy;
 import com.sun.enterprise.security.jmac.ConfigDomainParser;
 import com.sun.enterprise.security.jmac.WebServicesDelegate;
